@@ -10,21 +10,20 @@
     "@fortawesome/fontawesome-free": "^6.2.0",
     "bulma": "^0.9.4",
     "classnames": "^2.3.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "^4.0.3"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1"
   },
   "devDependencies": {
     "@cypress/webpack-dev-server": "^1.8.4",
-    "@mate-academy/cypress-tools": "^1.0.4",
-    "@mate-academy/eslint-config-react": "^0.0.11",
-    "@mate-academy/eslint-config-react-typescript": "^1.0.12",
-    "@mate-academy/scripts": "^1.2.12",
+    "@mate-academy/cypress-tools": "^1.0.5",
+    "@mate-academy/eslint-config-react-typescript": "^1.0.15",
+    "@mate-academy/scripts": "^1.7.0",
     "@mate-academy/students-ts-config": "*",
     "@mate-academy/stylelint-config": "^0.0.11",
-    "@types/node": "^17.0.23",
-    "@types/react": "^17.0.43",
-    "@types/react-dom": "^17.0.14",
+    "@types/node": "^16.18.80",
+    "@types/react": "^18.2.55",
+    "@types/react-dom": "^18.2.19",
     "cypress": "^9.5.3",
     "eslint": "^7.32.0",
     "eslint-plugin-cypress": "^2.11.2",
@@ -32,8 +31,9 @@
     "mochawesome": "^7.1.3",
     "mochawesome-merge": "^4.2.0",
     "mochawesome-report-generator": "^6.2.0",
-    "node-sass": "^6.0.1",
+    "node-sass": "^7.0.1",
     "postcss": "^8.4.12",
+    "prettier": "^3.2.5",
     "stylelint": "^15.11.0",
     "typescript": "^4.6.3"
   },
@@ -43,7 +43,8 @@
     "test": "mate-scripts test -l",
     "lint-js": "mate-scripts lint -j",
     "lint-css": "mate-scripts lint -s",
-    "lint": "npm run lint-js && npm run lint-css",
+    "format": "prettier --write './src/**/*.{ts,tsx}'",
+    "lint": "npm run format && npm run lint-js && npm run lint-css",
     "update": "mate-scripts update",
     "postinstall": "npm run update && cypress verify",
     "predeploy": "npm run build",
