@@ -1,7 +1,4 @@
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { App } from './App';
 
-ReactDOM.render(
-  <App />,
-  document.getElementById('root'),
-);
+createRoot(document.getElementById('root') as HTMLElement).render(<App />);
